+++ packages/zosfiles/__tests__/__system__/api/methods/hMigrate/HMigrate.system.test.ts
@@ -81,6 +81,27 @@ describe("Migrate Dataset", () => {
                 expect(listResponse.apiResponse.items[0].migr).toBe("YES");
                 expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigrationRequested.message);
             });
+            it("should migrate a sequential data set with response timeout", async () => {
+                let error;
+                let migrateResponse;
+                let listResponse;
+
+                try {
+                    migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet1, {responseTimeout: 5});
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet1, listOptions);
+                    Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+
+                expect(error).toBeFalsy();
+
+                expect(migrateResponse).toBeTruthy();
+                expect(migrateResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("YES");
+                expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigrationRequested.message);
+            });
             it("should migrate a sequential data set with wait = true", async () => {
                 const migrateOptions: IMigrateOptions = { "request": "hmigrate", "wait": true };
                 let error;
@@ -119,6 +140,27 @@ describe("Migrate Dataset", () => {
 
                 try {
                     migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet2);
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet2, listOptions);
+                    Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+
+                expect(error).toBeFalsy();
+
+                expect(migrateResponse).toBeTruthy();
+                expect(migrateResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("YES");
+                expect(migrateResponse.commandResponse).toContain(ZosFilesMessages.datasetMigrationRequested.message);
+            });
+            it("should migrate a partitioned dataset with response timeout", async () => {
+                let error;
+                let migrateResponse;
+                let listResponse;
+
+                try {
+                    migrateResponse = await HMigrate.dataSet(REAL_SESSION, dataSet2, {responseTimeout: 5});
                     listResponse = await List.dataSet(REAL_SESSION, dataSet2, listOptions);
                     Imperative.console.info(`Response: ${inspect(migrateResponse)}`);
                 } catch (err) {