+++ packages/zosfiles/__tests__/__system__/api/methods/hRecall/HRecall.system.test.ts
@@ -82,6 +82,27 @@ describe("Recall Dataset", () => {
                 expect(listResponse.apiResponse.items[0].migr).toBe("NO");
                 expect(recallResponse.commandResponse).toContain(ZosFilesMessages.datasetRecallRequested.message);
             });
+            it("should recall a sequential data set with response timeout", async () => {
+                let error;
+                let recallResponse;
+                let listResponse;
+
+                try {
+                    recallResponse = await HRecall.dataSet(REAL_SESSION, dataSet1, {responseTimeout: 5});
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet1, listOptions);
+                    Imperative.console.info(`Response: ${inspect(recallResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+
+                expect(error).toBeFalsy();
+
+                expect(recallResponse).toBeTruthy();
+                expect(recallResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("NO");
+                expect(recallResponse.commandResponse).toContain(ZosFilesMessages.datasetRecallRequested.message);
+            });
             it("should recall a sequential data set with wait = true", async () => {
                 const recallOptions: IRecallOptions = { "request": "hrecall", "wait": true };
                 let error;
@@ -121,6 +142,27 @@ describe("Recall Dataset", () => {
 
                 try {
                     recallResponse = await HRecall.dataSet(REAL_SESSION, dataSet2);
+                    listResponse = await List.dataSet(REAL_SESSION, dataSet2, listOptions);
+                    Imperative.console.info(`Response: ${inspect(recallResponse)}`);
+                } catch (err) {
+                    error = err;
+                    Imperative.console.info(`Error: ${inspect(err)}`);
+                }
+
+                expect(error).toBeFalsy();
+
+                expect(recallResponse).toBeTruthy();
+                expect(recallResponse.success).toBe(true);
+                expect(listResponse.apiResponse.items[0].migr).toBe("NO");
+                expect(recallResponse.commandResponse).toContain(ZosFilesMessages.datasetRecallRequested.message);
+            });
+            it("should recall a partitioned dataset with response timeout", async () => {
+                let error;
+                let recallResponse;
+                let listResponse;
+
+                try {
+                    recallResponse = await HRecall.dataSet(REAL_SESSION, dataSet2, {responseTimeout: 5});
                     listResponse = await List.dataSet(REAL_SESSION, dataSet2, listOptions);
                     Imperative.console.info(`Response: ${inspect(recallResponse)}`);
                 } catch (err) {