+++ packages/zosfiles/__tests__/__system__/cli/create/classicPds/__scripts__/command/command_create_classic_pds.sh
@@ -3,7 +3,7 @@ hlq=$1
 set -e
 
 echo "================Z/OS FILES CREATE DATA-SET-CLASSIC==============="
-zowe zos-files create data-set-classic "$1.test.data.set.classic"
+zowe zos-files create data-set-classic "$1.test.data.set.classic" $2
 if [ $? -gt 0 ]
 then
     exit $?