+++ packages/zosfiles/src/api/utils/ZosFilesUtils.ts
@@ -139,6 +139,10 @@ export class ZosFilesUtils {
             // do nothing
         }
 
+        if (options.responseTimeout != null) {
+            reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+        }
+
         return reqHeaders;
     }
 
@@ -252,6 +256,10 @@ export class ZosFilesUtils {
                 Headers.APPLICATION_JSON,
                 { "Content-Length": JSON.stringify(payload).length.toString() }
             ];
+
+            if (options.responseTimeout != null) {
+                headers.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+            }
 
             await ZosmfRestClient.putExpectString(session, endpoint, headers, payload);