+++ packages/zosfiles/src/cli/upload/ftds/FileToDataSet.handler.ts
@@ -35,7 +35,8 @@ export default class FileToDataSetHandler extends ZosFilesBaseHandler {
                 volume: commandParameters.arguments.volumeSerial,
                 binary: commandParameters.arguments.binary,
                 encoding: commandParameters.arguments.encoding,
-                task
+                task,
+                responseTimeout: commandParameters.arguments.responseTimeout
             });
 
         if (response.apiResponse) {